+++ src/Type/GiveawayWinners.php
@@ -28,5 +28,6 @@ final readonly class GiveawayWinners
         public ?true $onlyNewMembers = null,
         public ?true $wasRefunded = null,
         public ?string $prizeDescription = null,
+        public ?int $prizeStarCount = null,
     ) {}
 }