+++ src/Type/Payment/PaidMediaPurchased.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Type\Payment;
+
+use Vjik\TelegramBot\Api\Type\User;
+
+/**
+ * @see https://core.telegram.org/bots/api#paidmediapurchased
+ */
+final readonly class PaidMediaPurchased
+{
+    public function __construct(
+        public User $from,
+        public string $paidMediaPayload,
+    ) {}
+}