+++ tests/Client/PsrTelegramClientTest.php
@@ -141,7 +141,7 @@ final class PsrTelegramClientTest extends TestCase
                     /** @var Request $request */
                     $requestHeaders = $request->getHeaders();
                     $this->assertSame(['Content-Length', 'Content-Type'], array_keys($requestHeaders));
-                    $this->assertSame($requestHeaders['Content-Length'], ['390']);
+                    $this->assertSame($requestHeaders['Content-Length'], ['332']);
                     $this->assertSame([0], array_keys($requestHeaders['Content-Type']));
                     $this->assertSame(
                         1,
@@ -155,17 +155,14 @@ final class PsrTelegramClientTest extends TestCase
                         <<<TEXT
                             --$matches[1]
                             Content-Disposition: form-data; name="chat_id"
-                            Content-Length: 3
 
                             123
                             --$matches[1]
                             Content-Disposition: form-data; name="caption"
-                            Content-Length: 5
 
                             hello
                             --$matches[1]
                             Content-Disposition: form-data; name="photo"; filename="face.png"
-                            Content-Length: 14
                             Content-Type: image/png
 
                             test-file-body